refactor(BA-7820): keep the membership edge out of the entity catalog
Reverts the EntityMembershipEntityType declaration. The edge stays what
data/entity/KNOWLEDGE.md says the virtual-entity layer is — the layer that
expresses ownership, not an entity — so EntityMembershipID goes back to a plain
NewType in data/permission/id.py and GroupMembershipNotFoundError goes back to
ErrorDomain.GROUP on BackendAIError, listed again among the errors that stay
there for want of a declared row type.
