+++ src/ai/backend/manager/repositories/ops/v2/graph_write.py
@@ -18,10 +18,9 @@ from typing import Any
 import sqlalchemy as sa
 from sqlalchemy.dialects.postgresql import insert as pg_insert
 
-from ai.backend.common.data.entity.entity_membership import EntityMembershipID
 from ai.backend.common.data.entity.types import EntityIdentifier, EntityType
 from ai.backend.common.data.entity.virtual_entity import VirtualEntityID
-from ai.backend.common.data.permission.id import FieldPath
+from ai.backend.common.data.permission.id import EntityMembershipID, FieldPath
 from ai.backend.common.data.permission.types import Permission
 from ai.backend.manager.errors.permission import InvalidFieldPermission, VirtualEntityNotFound
 from ai.backend.manager.models.entity_label.row import EntityLabelRow
